refactor: move connection methods to publish connection
